Created Abstract Cache class and impl CacheBase interface and inherited all existing impl from the Abstract Cache.
